+++ chromatic/.storybook/main.ts
@@ -1,7 +1,4 @@
 import type { StorybookConfig } from '@storybook/web-components-vite';
-import { argv } from 'node:process';
-
-const devMode = !argv.includes('build');
 
 const config: StorybookConfig = {
   stories: [
@@ -11,7 +8,8 @@ const config: StorybookConfig = {
     '@storybook/addon-a11y', 
     '@storybook/addon-actions', 
     '@storybook/addon-storysource',
-    'storybook-addon-pseudo-states',
+    '@storybook/addon-themes',
+    'storybook-addon-pseudo-states',    
     {
       name: '@storybook/addon-essentials',
       options: {
@@ -29,7 +27,12 @@ const config: StorybookConfig = {
   },
   staticDirs: [
     { from: '../../packages/themes', to: '/themes' }
-  ]
+  ],
+  viteFinal: async config => {
+    const { mergeConfig } = await import('vite');
+
+    return mergeConfig(config, { logLevel: 'warn' });
+  }
 };
 
 export default config;