+++ packages/themes/sanoma-utbildning/index.ts
@@ -1,5 +1,4 @@
-import type { IconLibrary } from '@sl-design-system/icon';
-import { Icon } from '@sl-design-system/icon';
+import { Icon, type IconLibrary } from '@sl-design-system/icon';
 import { icons } from './icons.js';
 
 /**
@@ -7,5 +6,6 @@ import { icons } from './icons.js';
  * like icons, ...
  */
 export const setup = (): void => {
+  console.log('register Icons');
   Icon.register(icons as IconLibrary);
 };