+++ website/src/ts/components/status/status.ts
@@ -1,16 +1,15 @@
-import {LitElement, CSSResultGroup, type TemplateResult, html} from 'lit';
+import { BadgeSize } from '@sl-design-system/badge';
+import { LitElement, type TemplateResult, html } from 'lit';
 import { customElement, property } from 'lit/decorators.js';
-import {BadgeSize} from "@sl-design-system/badge";
 
 /** status `new` should be used only in the navigation. */
 export type StatusType = 'planned' | 'draft' | 'preview' | 'new' | 'stable' | 'deprecated';
 
 @customElement('ds-status')
 export class Status extends LitElement {
-
   @property() status: StatusType = 'planned';
 
-  @property() size: BadgeSize = '3xl';
+  @property() size: BadgeSize = 'lg';
 
   @property() version = '';
 
@@ -33,7 +32,9 @@ export class Status extends LitElement {
 
   override render(): TemplateResult {
     return html`
-                <sl-badge size=${this.size} variant=${this.badgeVariant}>${this.status ? this.status : 'planned'}</sl-badge>
+      <sl-badge emphasis="bold" size=${this.size} variant=${this.badgeVariant}
+        >${this.status ? this.status : 'planned'}</sl-badge
+      >
     `;
   }
 }